% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareQFeatures.R
\name{prepareQFeatures}
\alias{prepareQFeatures}
\title{Prepare a \code{QFeatures} object for .h5mu writing.}
\usage{
prepareQFeatures(object, sep = ":")
}
\arguments{
\item{object}{A \code{QFeatures} object.}

\item{sep}{Separator between the assay name and feature name.}
}
\value{
A prepared \code{QFeatures} object.
}
\description{
\code{prepareQFeatures()} makes feature row names globally unique across assays by
prefixing them with their assay name when needed.
}
